Fix inteface include propagation of modern CMake targets
Modern CMake interface dependencies are not added to the
INTERFACE_INCLUDE_DIRECTORIES property of a target,
but to INTERFACE_LINK_LIBRARIES as their included will
be propagated by CMake on compilation. However the package
config uses this variable to populate the include dirs,
which meant that coin and qt were missing from the path.

This patch adds the missing include directories (as well
as the GL include that was forgotten before).