Add a CMake option to disable the use of X11 libraries.
An underlying linking to X11 is currently performed, even on platforms
where it is not needed, like macOS.

That also prevents the portability of pre-compiled packages built on a
version of macOS to another macOS machine. If XQuartz (the macOS X11
package) is, in fact, installed on the build machine, the output library
is then silently linked to X11; then, X11 becomes a dependency of the
package on the client machine, even if not needed.

This is a breaker, for instance, when building Homebrew packages.

The option let users explicitely disable the use of X11.

----

A note: I'm not aware of potential uses of X11 linking on macOS. If no
use cases are foreseen, we could even disable the use of X11 completely
for SoQt on macOS, as it has been done for Coin:

    https://github.com/coin3d/coin/commit/58a1b4c3e968f96d3a2091fa5cb625f360ce6811